Run acceptance tests against dev control
This change enables running the acceptance tests against a tailnet managed by a real control (or devcontrol) server.

The acceptance tests run when the `-acc` flag is passed to `go test`. If `TEST_CONTROL_URL` and `TEST_API_KEY` are passed, then the test client and connector node will attempt to register with that control server and the test will modify the ACL of the tailnet that the API key belongs to in order to grant the client node the required DB connector app capability. If those environment variables are not set, the test will run against a fake control server.

Important notes
- The test currently requires someone to actively sit there and click on the auth URL to connect the connector and client tsnet servers to the example.com tailnet. In the future we might want to create auth keys via the API instead, so as to avoid that interactive part. 
- The tests overwrite the grant section of the tailnet's ACL and replaces it with just the one grant that the tests require to run. So another future improvement is to restore the original ACL properly during teardown to guarantee proper test isolation.

Updates tailscale.com/corp#34154

Signed-off-by: Gesa Stupperich <gesa@tailscale.com>